Header: show a Back button in place of About on app pages
The header is now a client component that swaps the left "About" pill
for a "Back" button when an app detail page is open (desktop included).
Back uses router.back() so the timeline scroll is restored, falling back
to "/" when there's no history. Removes the now-redundant in-content
BackLink component.
